created VKaplanQP, modified KaplanQP and PhononElectrode to accommodate
