+++ library/src/G4CMPPhononElectrode.cc
@@ -41,6 +41,7 @@
 // 20251115  G4CMP-539 -- Replace AddConstProperty() with UpdateMPT().
 // 20260212  G4CMP-581 -- Skip invalid phonons (null pointers), report skips.
 // 20260218  G4CMP-588 -- Fix change above by presenting secondary buffer.
+// 20260710  G4CMP-647 -- Replace G4CMPKaplanQP member with G4CMPVKaplanQP
 
 #include "G4CMPPhononElectrode.hh"
 #include "G4CMPGeometryUtils.hh"
@@ -50,6 +51,7 @@
 #include "G4CMPSurfaceProperty.hh"
 #include "G4CMPTrackUtils.hh"
 #include "G4CMPUtils.hh"
+#include "G4CMPVKaplanQP.hh"
 #include "G4Exception.hh"
 #include "G4ExceptionSeverity.hh"
 #include "G4LatticeManager.hh"
@@ -94,12 +96,17 @@ AbsorbAtElectrode(const G4Track& track, const G4Step& step,
 
   // Create KaplanQP simulator if not already available
   if (!kaplanQP) {
-    // Pass temperture through to KaplanQP if no already included
+    // Pass temperature through to KaplanQP if not already included
     if (!theSurfaceTable->ConstPropertyExists("temperature"))
       G4CMP::UpdateMPT(theSurfaceTable, "temperature",
 		       theLattice->GetTemperature());
 
-    kaplanQP = new G4CMPKaplanQP(theSurfaceTable, verboseLevel);
+    // not specifying identity of phonon-qp interaction model here
+    // if other interaction models are added, we will want to make a config
+    // parameter that will get passed in here so the user can specify
+    // the model to be used from a macro command
+    G4String interaction_model;
+    kaplanQP = getKaplanQP(interaction_model);
   }
 
   // Transfer phonon energy into superconducting film
@@ -138,9 +145,9 @@ ProcessAbsorption(const G4Track& track, const G4Step& step, G4double EDep,
   particleChange.ProposeNonIonizingEnergyDeposit(EDep);
 
   // Secondaries are emitted with cos(theta) distribution inward
-  //We'll need to eventually calculate the generalized surface norm
-  //based on the above surface norm and the track's velocity, so
-  //give track velocity
+  // We'll need to eventually calculate the generalized surface norm
+  // based on the above surface norm and the track's velocity, so
+  // give track velocity
   G4ThreeVector vDir = track.GetMomentumDirection();
   G4ThreeVector surfNorm = G4CMP::GetSurfaceNormal(step,vDir);
   
@@ -225,4 +232,15 @@ ProcessReflection(const G4Track& track, const G4Step& step,
 
   auto trackInfo = G4CMP::GetTrackInfo<G4CMPPhononTrackInfo>(track);
   trackInfo->IncrementReflectionCount();
+}
+
+// define phonon-qp interaction mechanism via string identifier; G4CMPKaplanQP default
+G4CMPVKaplanQP* G4CMPPhononElectrode::getKaplanQP(G4String identifier) const
+{
+  G4CMPVKaplanQP* val = nullptr;
+  if (identifier.empty() || identifier.find("aplan") != G4String::npos)
+  {
+    val = new G4CMPKaplanQP(theSurfaceTable, verboseLevel);
+  }
+  return val;
 }
